move profile login to the correct module
